tools: extract node creation
This commit prepares the structure of the tools directory for the
c-s frontend tests.

Changes:
- extracted scylla docker node creation logic to `tools/util/scylla_docker.py`
